Use aliases for GeoOrientation synonyms
